requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters
langgraph
langgraph-checkpoint-sqlite
faiss-cpu
sentence-transformers
huggingface-hub
